requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web framework (optional for web interface)
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
